refactor(front): change hooks folder structure
