+++ mylog/src/screens/feed/FeedDetailScreen.tsx
@@ -27,8 +27,8 @@ import PreviewImageList from '@/components/post/PreviewImageList';
 import FeedDetailHeader from '@/components/feed/FeedDetailHeader';
 import useLocationStore from '@/store/useLocationStore';
 import useDetailPostStore from '@/store/useDetailPostStore';
+import useModal from '@/hooks/useModal';
 import {useGetPost} from '@/hooks/queries/usePost';
-import useModal from '@/hooks/common/useModal';
 import {getDateLocaleFormat} from '@/utils/date';
 import {feedNavigations, mapNavigations} from '@/constants/navigations';
 import {colorHex, colors} from '@/constants/colors';
